# 02-openapi.yaml
openapi: 3.1.0
info:
  title: AI Second Brain API
  version: 1.1.1
  description: RAG backend — auth, ingest, hybrid search, ask (SSE + degraded mode), library, jobs, backups.
servers:
  - url: http://localhost:8000/v1
  - url: https://api.secondbrain.app/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable]
          additionalProperties: false
          properties:
            code:
              type: string
              enum: [VALIDATION, AUTH, ACCOUNT_LOCKED, NOT_FOUND, RATE_LIMITED, UPSTREAM_429, UPSTREAM_DEGRADED, PROCESSING, CONFLICT, INTERNAL]
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: string, format: uuid }
            retryable: { type: boolean }
    User:
      type: object
      required: [id, email, created_at]
      additionalProperties: false
      properties:
        id: { type: string }
        email: { type: string, format: email }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
    RegisterRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 72 }
        display_name: { type: string, maxLength: 80 }
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, maxLength: 72 }
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      additionalProperties: false
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, description: access lifetime seconds }
        user: { $ref: '#/components/schemas/User' }
    RefreshRequest:
      type: object
      required: [refresh_token]
      additionalProperties: false
      properties:
        refresh_token: { type: string, minLength: 20 }
    LogoutRequest:
      type: object
      additionalProperties: false
      properties:
        refresh_token: { type: string, minLength: 20 }
paths:
  /healthz:
    get:
      security: []
      operationId: health
      responses:
        '200': { description: ok }
  /readyz:
    get:
      security: []
      operationId: ready
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready: { type: boolean }
                  models: { type: object }
                  sweep: { type: object }
                  backup: { type: object }
  /auth/register:
    post:
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '400':
          description: error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: email taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          description: rate limited
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: invalid credentials (identical for bad email/password)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          description: locked or rate limited
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /auth/refresh:
    post:
      security: []
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: rotated pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: invalid/reused (reuse revokes chain)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /auth/logout:
    post:
      operationId: logout
      parameters:
        - name: all
          in: query
          schema: { type: boolean, default: false }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '200': { description: revoked }
  /auth/me:
    get:
      operationId: me
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          description: unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /admin/backups:
    get:
      security: []
      operationId: listBackups
      parameters:
        - name: X-Restore-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        '200': { description: snapshot catalog }
        '401':
          description: unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /admin/backup:
    post:
      security: []
      operationId: enqueueBackup
      parameters:
        - name: X-Restore-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        '202': { description: backup job queued }
  /admin/restore:
    post:
      security: []
      operationId: restoreBackup
      parameters:
        - name: X-Restore-Token
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                backup_id: { type: string }
      responses:
        '200': { description: restored }
        '409':
          description: users table not empty
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /admin/users:
    get:
      operationId: adminListUsers
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        '200': { description: operator user list }
        '404': { description: not operator }
  /admin/users/{user_id}/unlock:
    post:
      operationId: adminUnlockUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: unlocked }
        '404': { description: not found }
  /admin/users/{user_id}/revoke-sessions:
    post:
      operationId: adminRevokeSessions
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: refresh chain revoked }
        '404': { description: not found }
  /admin/users/{user_id}/send-reset:
    post:
      operationId: adminSendReset
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: reset email attempted }
        '404': { description: not found }
  /admin/users/{user_id}:
    delete:
      operationId: adminDeleteUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: deleted }
        '404': { description: not found }
        '409': { description: cannot delete self }
  /documents:
    get:
      operationId: listDocuments
      responses:
        '200': { description: ok }
        '401':
          description: unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }